requirements.txt
chromadb==1.3.5
fastapi==0.124.0
langchain==0.0.208
langchain_core==1.1.2
langchain_openai==1.1.1
langgraph==1.0.4
openai==2.9.0
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
sentence_transformers==5.1.2
pyarrow==22.0.0
fastparquet==2024.11.0
uvicorn[standard]==0.23.2
streamlit==1.30.0
python-multipart==0.0.6
